+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add buffer relocation in reader
+
 ## [2.4.0] - 2024-03-27
 
 This is a feature release of **picocsv**.